x86/tdx: Move virt exception handler code to virt_exception.c
The TDX using virt exception to handle mmio opeartion, since virt
exception is a cpu feature, it can also being used by other cc_platform.
So move the common virt exception handler code to virt_exception.c
